Create separate file for the custom level editor tab.
